refactor: Move utility functions to utils directory
- Moved `check_dependencies` to `utils/system.py`
- Moved `convert_to_mkv` to `utils/video.py`
- Moved settings-related functions (`load_settings`, `save_settings`) into the `Config` class in `utils/Config.py`
- Cleaned up `dl-kp.py` to use the new refactored code.

This refactoring improves the code structure and separation of concerns.
